create weather component.refactor code accordindly.finish user story 4
